requirements.txt
# PDF Parsing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# LLM Integration
openai>=1.40.0
anthropic>=0.34.0

# Semantic Similarity
sentence-transformers>=3.0.0

# Template & Export
Jinja2>=3.1.4
weasyprint>=62.0

# Web Interface
streamlit>=1.38.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.8.0
loguru>=0.7.2
tenacity>=9.0.0
